Organized files by using folders. Minor changes.
