+++ NOTEA/Helpers/EnumHelper.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations;
 using System.ComponentModel;
 using System.Reflection;
 
-namespace NOTEA.Models
+namespace NOTEA.Helpers
 {
     public static class EnumHelper<T>
     {
