+++ NOTEA/Helpers/NameValidator.cs
@@ -1,10 +1,10 @@
-﻿namespace NOTEA.Models
+﻿namespace NOTEA.Helpers
 {
     public static class NameValidator
     {
         public static bool IsValidFilename(this string name)
         {
-            if(name == null || name.Length > 80 || name[0] == '.' || name.Contains('\\') ||
+            if (name == null || name.Length > 80 || name[0] == '.' || name.Contains('\\') ||
                 name.Contains('/') || name.Contains('?') || name.Contains('*') ||
                 name.Contains(':') || name.Contains('"') || name.Contains('<') ||
                 name.Contains('>') || name.Contains('|'))
