+++ NOTEA/Models/ConspectModels/ConspectListModel.cs
@@ -1,12 +1,12 @@
 ﻿using System.Collections;
 
-namespace NOTEA.Models
+namespace NOTEA.Models.ConspectModels
 {
     public class ConspectListModel<ConspectType>
     {
         public List<ConspectType> Conspects = new List<ConspectType>();
         public ConspectListModel() { }
-        public ConspectListModel(List<ConspectType> conspects) { this.Conspects = conspects; }
+        public ConspectListModel(List<ConspectType> conspects) { Conspects = conspects; }
 
     }
