--- NOTEA/Models/RecordModel.cs
@@ -1,7 +1,0 @@
-﻿namespace NOTEA.Models
-{
-    public record RecordModel
-    {
-        public DateTime Date {  get; set; }
-    }
-}
