+++ NOTEA/Models/RecordModels/RecordModel.cs
@@ -1,0 +1,7 @@
+﻿namespace NOTEA.Models.RecordModels
+{
+    public record RecordModel
+    {
+        public DateTime Date { get; set; }
+    }
+}
