+++ NOTEA/Models/UserModels/UserListModel.cs
@@ -1,4 +1,4 @@
-﻿namespace NOTEA.Models
+﻿namespace NOTEA.Models.UserModels
 {
     public class UserListModel
     {
