+++ NOTEA/Models/UserModels/UserModel.cs
@@ -1,11 +1,11 @@
-﻿namespace NOTEA.Models
+﻿namespace NOTEA.Models.UserModels
 {
     public class UserModel
     {
         public string Username { get; set; }
         public string Password { get; set; }
         public string PasswordCheck { get; set; }
-        public UserModel (string username, string password)
+        public UserModel(string username, string password)
         {
             Username = username;
             Password = password;
