+++ NOTEA/Views/Conspect/CreateConspects.cshtml
@@ -1,5 +1,7 @@
 ﻿@inject IHttpContextAccessor context
-@model NOTEA.Models.ConspectModel
+@using NOTEA.Models.ConspectModels;
+@using NOTEA.Helpers;
+@model ConspectModel
 
 @{
     ViewData["Title"] = "Create conspects";
