+++ NOTEA/Views/Conspect/UploadConspect.cshtml
@@ -1,5 +1,6 @@
 ﻿@inject IHttpContextAccessor context
-@model NOTEA.Models.FileHandlerModel
+@using NOTEA.Models.FileModels
+@model FileHandlerModel
 
 @{
     ViewData["Title"] = "Conspect upload";
