+++ NOTEA/Views/Conspect/ViewConspect.cshtml
@@ -1,5 +1,6 @@
 ﻿@inject IHttpContextAccessor context
-@model NOTEA.Models.ConspectModel
+@using NOTEA.Models.ConspectModels;
+@model ConspectModel
 
 @{
     ViewData["Title"] = "ViewConspect";
