+++ NOTEA/Views/Shared/Error.cshtml
@@ -1,4 +1,5 @@
-﻿@model ErrorViewModel
+﻿@using NOTEA.Models.ErrorModels;
+@model ErrorViewModel
 @{
     ViewData["Title"] = "Error";
 }
