+++ NOTEA/Views/User/LogIn.cshtml
@@ -1,5 +1,6 @@
 ﻿@inject IHttpContextAccessor context
-@model NOTEA.Models.UserModel
+@using NOTEA.Models.UserModels;
+@model UserModel
 
 @{
     ViewData["Title"] = "Log In";
