+++ NOTEA/Views/User/SignIn.cshtml
@@ -1,5 +1,6 @@
 ﻿@inject IHttpContextAccessor context
-@model NOTEA.Models.UserModel
+@using NOTEA.Models.UserModels;
+@model UserModel
 
 @{
     ViewData["Title"] = "Sign In";
